chore: moves `actions` file to `/dashboard` folder
